made world.html, url in urls.py, dicts in views.py
Added the dictionary values in views.py which also sets the variables to be used in world.html. It should be changed depending on which direction you choose.
title.html has a start button that doesn't work, but it should take you to the first world page.
Added a url path in urls.py